requirements.txt
open_clip_torch
transformers
numpy
pandas
beir
openpyxl
datasets